+++ praw/models/reddit/live.py
@@ -173,6 +173,44 @@ class LiveContributorRelationship(object):
                 'permissions': self._handle_permissions(permissions)}
         self.thread._reddit.post(url, data=data)
 
+    def update_invite(self, redditor, permissions=None):
+        """Update the contributor invite permissions for ``redditor``.
+
+        :param redditor: A redditor name (e.g., ``'spez'``) or
+            :class:`~.Redditor` instance.
+        :param permissions: When provided (not ``None``), permissions should
+            be a list of strings specifying which subset of permissions to
+            grant (other permissions are removed). An empty list ``[]``
+            indicates no permissions, and when not provided (``None``),
+            indicates full permissions.
+
+        For example, to set all permissions to the invitation, try:
+
+        .. code:: python
+
+           thread = reddit.live('ukaeu1ik4sw5')
+           thread.contributor.update_invite('spez')
+
+        To set 'access' and 'edit' permissions (and to remove other
+        permissions) to the invitation, try:
+
+        .. code:: python
+
+           thread.contributor.update_invite('spez', ['access', 'edit'])
+
+        To remove all permissions from the invitation, try:
+
+        .. code:: python
+
+           thread.contributor.update_invite('spez', [])
+
+        """
+        url = API_PATH['live_update_perms'].format(id=self.thread.id)
+        data = {'name': str(redditor),
+                'type': 'liveupdate_contributor_invite',
+                'permissions': self._handle_permissions(permissions)}
+        self.thread._reddit.post(url, data=data)
+
 
 class LiveThread(RedditBase):
     """An individual LiveThread object."""